fix: 검증용 실행 api 세부 개선 사항 3건 반영

+++ judge-service/src/main/java/com/maesamco/judge/application/command/ExecutionValidationTestCase.java
@@ -3,7 +3,7 @@ package com.maesamco.judge.application.command;
 public record ExecutionValidationTestCase(
         String input,
         String expectedOutput,
-        int cpuTimeLimitSeconds,
+        double cpuTimeLimitSeconds,
         int memoryLimitKb
 ) {
 }
